Move search feature into its own sub-app
This continues on the folder structure work to move features
into their own express sub-apps to keep concerns neatly separated.
