+++ src/apps/search/index.js
@@ -1,0 +1,13 @@
+const router = require('./router')
+const searchController = require('./search.controller')
+const searchService = require('./search.service')
+
+module.exports = {
+  router,
+  controllers: {
+    search: searchController,
+  },
+  services: {
+    search: searchService,
+  },
+}
