+++ src/controllers/company-add.controller.js
@@ -2,7 +2,7 @@ const express = require('express')
 const { ukOtherCompanyOptions, foreignOtherCompanyOptions } = require('../options')
 const { getCHCompany } = require('../repos/company.repo')
 const { isBlank, toQueryString } = require('../lib/controller-utils')
-const { searchLimitedCompanies } = require('../services/search.service')
+const { searchLimitedCompanies } = require('../apps/search/search.service')
 const { getDisplayCH, getDisplayCompany } = require('../services/company-formatting.service')
 const { getPagination } = require('../lib/pagination')
 const { companyDetailsLabels, chDetailsLabels, companyTypeOptions } = require('../labels/company-labels')
